+++ ui/src/components/dashboard/composables/useDashboard.ts
@@ -1,6 +1,6 @@
-import Utils from "../../utils/utils";
+import Utils from "../../../utils/utils.js";
 import {cssVariable} from "@kestra-io/ui-libs";
-import {getConsistentHEXColor} from "../../utils/charts.js";
+import {getConsistentHEXColor} from "../../../utils/charts.js";
 
 const getOrCreateLegendList = (chart, id, direction = "row", width = "100%") => {
     const legendContainer = document.getElementById(id);
